refactor: move correctness calculation to a separate file
